feat [web]: Overhaul adjustments accordion with dithering support
